Harness: one structural equality for both comparison paths
The sandbox baseline gate compared values by serializing both sides with
stableStringify, while the argTypes comparator already had a structural
deepEqual. Same question, and the serializing one was the weaker answer: it
allocated a full sorted clone plus a pretty-printed string per field per
component only to throw both away, and it collapsed distinctions a byte-exact
gate should keep.

deepEqual moves to its own module and both paths use it. stableStringify follows
the write path it actually serves, where sorted pretty JSON is the point.

Key-order insensitivity was the property the serializing form provided and
nothing pinned it, so a test now does - baselines are written key-sorted while a
build emits its own order, so losing it would report drift on every verify run.

Also names the collision between the gate's UNRESOLVED_STUBS and the perf
engine's RESOLVED_TYPES, where `undefined` means opposite things, and renames the
sandbox recorder's readCommitted, which read a directory of baselines rather than
the snapshot text the identically-named Angular helper reads.
