+++ code/lib/docgen-harness/src/compare/deep-equal.ts
@@ -1,0 +1,35 @@
+/**
+ * Structural equality, insensitive to key order.
+ *
+ * Short-circuits on the first difference and allocates nothing, which is why the comparison paths use
+ * it rather than comparing two serializations of the same values. It also keeps `{ a: undefined }`
+ * distinct from `{}`, where a JSON round-trip would collapse them.
+ */
+export function deepEqual(a: unknown, b: unknown): boolean {
+  if (Object.is(a, b)) {
+    return true;
+  }
+  if (Array.isArray(a) && Array.isArray(b)) {
+    return a.length === b.length && a.every((item, index) => deepEqual(item, b[index]));
+  }
+  if (
+    a !== null &&
+    b !== null &&
+    typeof a === 'object' &&
+    typeof b === 'object' &&
+    !Array.isArray(a) &&
+    !Array.isArray(b)
+  ) {
+    const aKeys = Object.keys(a);
+    const bKeys = Object.keys(b);
+    return (
+      aKeys.length === bKeys.length &&
+      aKeys.every(
+        (key) =>
+          Object.prototype.hasOwnProperty.call(b, key) &&
+          deepEqual((a as Record<string, unknown>)[key], (b as Record<string, unknown>)[key])
+      )
+    );
+  }
+  return false;
+}
