+++ code/lib/docgen-harness/src/sandbox-baselines/compare-baselines.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from 'vitest';
 
-import { compareBaselines, formatFindings, stableStringify } from './compare-baselines.ts';
+import { compareBaselines, formatFindings } from './compare-baselines.ts';
+import { stableStringify } from './stable-stringify.ts';
 import type { SandboxBaseline } from './read-static-docgen.ts';
 
 type ArgTypes = NonNullable<SandboxBaseline['argTypes']>;
@@ -29,6 +30,19 @@ const undocumented = (overrides: Partial<SandboxBaseline> = {}): SandboxBaseline
 describe('compareBaselines', () => {
   it('reports nothing when both sides agree', () => {
     expect(compareBaselines({ button: documented() }, { button: documented() })).toEqual([]);
+  });
+
+  it('reports nothing when the two sides spell the same content in a different key order', () => {
+    // Baselines are written key-sorted but a build emits its own order, so a key-order-sensitive
+    // comparison would report drift on every verify run.
+    const sorted = { name: 'label', table: { category: 'inputs' }, type: { name: 'string' } };
+    const shuffled = { type: { name: 'string' }, name: 'label', table: { category: 'inputs' } };
+    expect(
+      compareBaselines(
+        { button: documented({ argTypes: { label: sorted } as ArgTypes }) },
+        { button: documented({ argTypes: { label: shuffled } as ArgTypes }) }
+      )
+    ).toEqual([]);
   });
 
   it('flags a component that stopped being documented as a regression', () => {
